The Game
Here's the game state as it is. There's some controller support. Menu is
semi there. Inventory system is abysmal, but the hard work is mostly
done. Movement is not smooth and the player is controlled via state
machine. Parallax background scrolling is there. Yada yada.

Let me know if you have questions.