New branch.
So this one has jumping and z_space functionality, you can pick up 5
items, and it'll display them on a hotbar, though the functionality is
not fully there.

The inventory system is extremely and could even be used to hold data
not exclusive to items and weapons.

Oh and jumping. Yeah, that's rad.

I'll be branching off from here to see if I can maybe make a heightmap
instead. If I can, this game is going to be swagtacular.